bug fixes, improvements

+++ src/main/resources/static/scripts/form-crud-ajax.js
@@ -1,6 +1,4 @@
 $(document).ready(function() {
-    let isEditing = false;
-
     // Include CSRF token in the headers (for Spring Security)
     // let csrfToken = $("meta[name='_csrf']").attr("content");
     // let csrfHeader = $("meta[name='_csrf_header']").attr("content");
@@ -9,6 +7,8 @@ $(document).ready(function() {
     // headers[csrfHeader] = csrfToken;
     headers["Content-Type"] = "application/json";
     // headers["Authorization"] = "Bearer " + jwtToken;
+
+    const filterSelect = $('#filterSelect');
 
     const taskFormModal = $('#taskFormModal');
     const taskModalLabel = $('#taskModalLabel');
@@ -24,18 +24,58 @@ $(document).ready(function() {
     const taskCategories = $('#taskCategories');
     const taskPriority = $('#taskPriority');
     const taskSubmitBtn = $('#taskSubmitBtn');
+    const taskDeleteBtn = $('#taskDeleteBtn');
 
+
+    // CATEGORIES
     $.ajax({
         url: '/categories',
         type: 'GET',
         success: function(categories) {
-            categories.forEach(category => {
+            const categoryData = categories.map(category => {
                 taskCategories.append(new Option(category, category, false, false));
+                return { id: category, text: category, group: 'Categories'};
             });
+
+            initializeFilter(categoryData, filterSelect);
         },
         error: function(xhr) {
             console.error('Error loading categories:', xhr);
         }
+    });
+
+    $('#addCategory').on('click', function() {
+        $('#categoryModal').modal('show');
+    });
+
+    $('#categoryForm').submit(function(e) {
+        e.preventDefault();
+        const category = $('#categoryName').val();
+        $.ajax({
+            url: '/categories',
+            type: 'POST',
+            data: { name: category },
+            success: function() {
+                taskCategories.append(new Option(category, category, false, true));
+            },
+            error: function(xhr) {
+                console.error('Error adding category:', xhr);
+            }
+        });
+    });
+
+    taskCategories.select2({
+        placeholder: 'Select categories',
+        theme: "bootstrap-5",
+        allowClear: true
+    }).on('select2:select select2:unselect', function(e) {
+        updatePlaceholder(taskCategories, "Select categories", "Search");
+    });
+
+    taskPriority.select2({
+        theme: "bootstrap-5",
+        placeholder: 'Select priority',
+        dropdownParent: taskPriority.parent()
     });
 
 
@@ -50,20 +90,11 @@ $(document).ready(function() {
         taskDateCreated.val(new Date().toISOString().substring(0, 10)); // Date in YYYY-MM-DD format
         taskDueDate.val('');
         taskCategories.val(null).trigger('change');
-        taskPriority.val('');
+        taskPriority.val(null).trigger('change');
+        taskDeleteBtn.css('display', 'none');
         $('.error').text('');
 
         taskFormModal.modal('show');
-    });
-
-    $('#addCategory').on('click', function() {
-        $('#categoryModal').modal('show');
-    });
-
-    taskCategories.select2({
-        theme: "bootstrap-5",
-        closeOnSelect: false,
-        allowClear: true
     });
 
 
@@ -80,32 +111,30 @@ $(document).ready(function() {
             descCharCount.text(task.description.length + ' / ' + taskDescription.attr('maxlength'));
             taskDateCreated.val(task.dateCreated);
             taskDueDate.val(task.dueDate);
-
-            // task.categories.forEach(category => {
-            //     taskCategories.append(new Option(category, category, false, true));
-            // });
             taskCategories.val(task.categories).trigger('change');
-
-
-            taskPriority.val(task.priority);
-            taskFormModal.modal('show');
+            taskPriority.val(task.priority).trigger('change');
+            taskDeleteBtn.css('display', 'inline-block');
             $('.error').text('');
-            isEditing = true;
+            taskFormModal.modal('show');
         });
     });
 
-    // DELETING
-    table.on('click', '.deleteBtn', function() {
-        const id = $(this).data('id');
-        $.ajax({
-            url: '/tasks/' + id,
-            type: 'DELETE',
-            headers: headers,
-            success: function() {
-                $('.task-row[data-id="' + id + '"]').remove();
-            }
-        });
+    taskDeleteBtn.click(function() {
+        if (confirm('Are you sure you want to delete this task?')) {
+            const id = taskId.val();
+
+            $.ajax({
+                url: '/tasks/' + id,
+                type: 'DELETE',
+                headers: headers,
+                success: function() {
+                    $('.task-row[data-id="' + id + '"]').remove();
+                    taskFormModal.modal('hide');
+                }
+            });
+        }
     });
+
 
     // SAVING
     $('#taskForm').submit(function(e) {
@@ -120,8 +149,6 @@ $(document).ready(function() {
         });
         formDataDict['categories'] = taskCategories.val();
 
-        console.log(formDataDict);
-
         $.ajax({
             url: '/tasks',
             type: 'POST',
@@ -131,12 +158,12 @@ $(document).ready(function() {
                 formDataDict['id'] = task.id; // ID specified by the server
                 updateOrAddTableRow(formDataDict);
                 taskFormModal.modal('hide');
-                isEditing = false;
             },
             error: function(xhr) {
                 const errors = xhr.responseJSON;
                 $('.error').text('');
                 for (const field in errors) {
+                    console.log(field);
                     $('#' + field + 'Error').text(errors[field]);
                 }
             },
@@ -161,12 +188,37 @@ $(document).ready(function() {
     $('th[data-sort]').on('click', function() {
         const column = $(this).data('sort');
         const order = table.data('order') === 'asc' ? 'desc' : 'asc';
+        const filters = getFilters(filterSelect);
+        // const icon = order === 'asc' ? '▲' : '▼';
         table.data('order', order);
+
+        console.log(filters.categories);
+        console.log(filters.priorities);
 
         $.ajax({
             url: '/tasks/sort',
             method: 'GET',
-            data: { sort: column, order: order },
+            data: { sort: column, order: order, categories: filters.categories, priorities: filters.priorities},
+            traditional: true,
+            success: function(data) {
+                let rows = '';
+                $.each(data, function(index, task) {
+                    rows += createNewRow(task).prop('outerHTML');
+                });
+
+                $('#taskTable tbody').html(rows);
+            }
+        });
+    });
+
+    // FILTERING
+    filterSelect.on('change', function() {
+        const filters = getFilters(filterSelect);
+
+        $.ajax({
+            url: '/tasks/filter',
+            method: 'GET',
+            data: { categories: filters.categories, priorities: filters.priorities },
             success: function(data) {
                 let rows = '';
                 $.each(data, function(index, task) {
@@ -176,24 +228,32 @@ $(document).ready(function() {
             }
         });
     });
-    //
-    // $(document).on('mousedown', function(event) {
-    //     console.log(taskCategories.data('select2').isOpen());
-    //
-    //     const target = $(event.target);
-    //     // If the target is outside the modal and the dropdown is open, close the dropdown
-    //     if (!target.closest('#taskFormModalInner').length && !target.closest('.select2-container').length) {
-    //         console.log("EXECUTED");
-    //         event.stopPropagation();
-    //         event.stopImmediatePropagation();
-    //         if (taskCategories.data('select2').isOpen()) {
-    //             console.log("CLOSING");
-    //             taskCategories.select2('close');
-    //         }
-    //     }
-    // });
+
+    // SEARCHING
+    $('#searchTerm').keyup(function() {
+        const searchTerm = $(this).val();
+        // console.log(searchTerm);
+        $.ajax({
+            url: '/tasks/search',
+            type: 'GET',
+            data: { term: searchTerm },
+            success: function(data) {
+                let rows = '';
+                $.each(data, function(index, task) {
+                    rows += createNewRow(task).prop('outerHTML');
+                });
+                $('#taskTable tbody').html(rows);
+            }
+        });
+    });
 });
 
+// --------- HELPER FUNCTIONS
+
+function clearCategoryModal() {
+    $('#categoryName').val('');
+    $('.error').text('');
+}
 
 function updateCharacterCount() {
     let textarea = $("#taskDescription");
@@ -208,6 +268,25 @@ function createNewRow(task) {
     newRow.append('<td>' + task.priority +'</td>');
     newRow.append('</tr>');
 
+    const dueDate = new Date(task.dueDate);
+    const now = new Date();
+    const oneDay = 24 * 60 * 60 * 1000;
+    const daysUntilDue = Math.ceil(Math.abs((dueDate - now) / oneDay));
+
+    console.log(dueDate);
+    console.log(now);
+    console.log(daysUntilDue);
+    console.log(dueDate - now);
+    console.log((dueDate - now) / oneDay);
+
+    if (daysUntilDue <= 1) {
+        newRow.addClass('due-in-one-day');
+    } else if (daysUntilDue === 2) {
+        newRow.addClass('due-in-two-days');
+    }
+
+
+
     return newRow;
 }
 
@@ -216,4 +295,57 @@ function updateRow(existingRow, task) {
     existingRow.find('td:eq(1)').text(task.dueDate);
     existingRow.find('td:eq(2)').text(task.categories);
     existingRow.find('td:eq(3)').text(task.priority);
+}
+
+function initializeFilter(categories, $element) {
+    $element.select2({
+        placeholder: 'Select filters',
+        closeOnSelect: false,
+        allowClear: true,
+        theme: 'bootstrap-5',
+        data: [
+            {
+                text: 'Categories',
+                children: categories
+            },
+            {
+                text: 'Priorities',
+                children: [
+                    { id: 'High', text: 'High', group: 'Priorities'},
+                    { id: 'Medium', text: 'Medium', group: 'Priorities' },
+                    { id: 'Low', text: 'Low', group: 'Priorities' }
+                ]
+            }
+        ]
+    }).on('select2:select select2:unselect', function(e) {
+        updatePlaceholder($element, "Select filters", "Search");
+    });
+}
+
+function updatePlaceholder($element, msgDefault, msgSelected) {
+    const $searchField = $element.data('select2').dropdown.$search || $element.data('select2').selection.$search;
+    const selectedItems = $element.select2('data');
+    if (selectedItems.length > 0) {
+        $searchField.attr('placeholder', msgSelected);
+        $searchField.css('width', '100%');
+    } else {
+        $searchField.attr('placeholder', msgDefault);
+    }
+}
+
+function getFilters(filterSelect) {
+    const selectedCategories = [];
+    const selectedPriorities = [];
+
+    filterSelect.data('select2').dataAdapter.current(function(data) {
+        data.forEach(group => {
+            if (group.group === 'Categories') {
+                selectedCategories.push(group.id);
+            } else if (group.group === 'Priorities') {
+                selectedPriorities.push(group.id);
+            }
+        });
+    });
+
+    return { categories: selectedCategories, priorities: selectedPriorities };
 }
